+++ Contents/Code/update_tools.py
@@ -1,21 +1,199 @@
 # Import internal tools
 from logging import Logging
+from region_tools import RegionTool
 import re
 
 # Setup logger
 log = Logging()
 
 
-class AlbumUpdateTool:
-    UPDATE_URL = 'https://api.audnex.us/books/'
-
-    def __init__(self, force, lang, media, metadata):
-        self.date = None
+class UpdateTool:
+    def __init__(self, content_type, force, lang, media, metadata, prefs):
+        self.content_type = content_type
         self.force = force
-        self.genres = None
         self.lang = lang
         self.media = media
         self.metadata = metadata
+        self.prefs = prefs
+        self.region = self.extract_region_from_id()
+
+    def build_url(self):
+        """
+            Builds the URL for the API request.
+        """
+        # Get the current region
+        self.region_override = self.region if self.region else self.prefs['region']
+        # Set the region helper
+        region_helper = RegionTool(
+            region=self.region_override, content_type=self.content_type, id=self.extract_asin_from_id())
+
+        update_url = region_helper.get_id_url()
+        log.debug('Update URL: ' + update_url)
+        return update_url
+
+    def collect_metadata_to_log(self):
+        # Start an array with common metadata
+        data_to_log = [{'ASIN': self.metadata.id}]
+
+        # Determine which metadata to log
+        if self.content_type == 'books':
+            data_to_log.extend(
+                [
+                    {'Book poster URL': self.thumb},
+                    {'Book publisher': self.metadata.studio},
+                    {'Book release date': str(
+                        self.metadata.originally_available_at)},
+                    {'Book sort title': self.metadata.title_sort},
+                    {'Book summary': self.metadata.summary},
+                    {'Book title': self.metadata.title},
+                ]
+            )
+        elif self.content_type == 'authors':
+            data_to_log.extend(
+                [
+                    {'Author bio': self.metadata.summary},
+                    {'Author name': self.metadata.title},
+                    {'Author poster URL': self.thumb},
+                    {'Author sort name': self.metadata.title_sort},
+                ]
+            )
+
+        return data_to_log
+
+    def collect_metadata_arrs_to_log(self):
+        # Start an array with common metadata
+        multi_arr = [{'Genres': self.metadata.genres}]
+
+        # Book metadata
+        book_multi_arr = [
+            {'Moods (Authors)': self.metadata.moods},
+            {'Styles (Narrators)': self.metadata.styles},
+        ]
+
+        # Determine which metadata to log
+        if self.content_type == 'book':
+            multi_arr.extend(book_multi_arr)
+
+        return multi_arr
+
+    def extract_asin_from_id(self):
+        """
+            Extracts the ASIN from the ID.
+        """
+        # Get the ASIN from the ID
+        asin = self.metadata.id.split('_')[0]
+        log.debug('Extracted ASIN from ID: ' + asin)
+        return asin
+
+    def extract_region_from_id(self):
+        """
+            Extracts the region from the ASIN and sets the region.
+        """
+        # Get the region from the ASIN
+        try:
+            region = self.metadata.id.split('_')[1]
+            log.debug('Extracted region from ASIN: ' + region)
+        except IndexError:
+            log.info('No region found in ID, using default region.')
+            region = 'us'
+            # Save the region to the ID
+            self.metadata.id = self.metadata.id + '_' + region
+        # Set region and ASIN
+        return region
+
+    # Writes metadata information to log.
+    def log_update_metadata(self):
+        # Send a separator to the log
+        log.separator(
+            msg=(
+                'FINALIZED: ' + self.metadata.title +
+                ', ID: ' + self.metadata.id
+            ),
+            log_level="info"
+        )
+
+        # Collect metadata to log
+        data_to_log = self.collect_metadata_to_log()
+        log.metadata(data_to_log, log_level="info")
+
+        # Collect metadata arrays to log
+        multi_arr = self.collect_metadata_arrs_to_log()
+        log.metadata_arrs(multi_arr, log_level="info")
+
+        log.separator(log_level="info")
+
+
+class AlbumUpdateTool(UpdateTool):
+    def set_date(self):
+        """
+            Sets the date.
+        """
+        if self.date is not None:
+            if not self.metadata.originally_available_at or self.force:
+                self.metadata.originally_available_at = self.date
+
+    def set_rating(self):
+        """
+            Sets the rating.
+        """
+        # We always want to refresh the rating
+        if self.rating:
+            self.metadata.rating = float(self.rating) * 2
+
+    def set_summary(self):
+        """
+            Sets the summary.
+        """
+        if not self.metadata.summary or self.force:
+            self.metadata.summary = self.synopsis
+
+    def set_studio(self):
+        """
+            Sets the studio.
+        """
+        if not self.metadata.studio or self.force:
+            self.metadata.studio = self.studio
+
+    def set_sort_title(self):
+        """
+            Sets the sort title.
+        """
+        # Add series/volume to sort title where possible.
+        series_with_volume = ''
+        if self.series and self.volume:
+            series_with_volume = self.series + ', ' + self.volume
+        # Only include subtitle in sort if not in a series
+        if not self.volume:
+            self.title = self.metadata.title
+        if not self.metadata.title_sort or self.force:
+            self.metadata.title_sort = ' - '.join(
+                filter(
+                    None, [(series_with_volume), self.title]
+                )
+            )
+
+    def set_title(self):
+        """
+            Sets the title.
+        """
+        # If the `simplify_title` option is selected, don't append subtitle
+        # and remove extra endings on the title
+        if self.prefs['simplify_title']:
+            album_title = self.simplify_title()
+        elif self.subtitle:
+            album_title = self.title + ': ' + self.subtitle
+        else:
+            album_title = self.title
+        if not self.metadata.title or self.force:
+            self.metadata.title = album_title
+
+    def parse_api_response(self, response):
+        """
+            Parses keys from API into helper variables if they exist.
+        """
+        # Set empty variables
+        self.date = None
+        self.genres = None
         self.rating = None
         self.series = ''
         self.series2 = ''
@@ -24,10 +202,6 @@ class AlbumUpdateTool:
         self.volume = ''
         self.volume2 = ''
 
-    def parse_api_response(self, response):
-        """
-            Parses keys from API into helper variables if they exist.
-        """
         if 'authors' in response:
             self.author = response['authors']
         if 'releaseDate' in response:
@@ -68,38 +242,6 @@ class AlbumUpdateTool:
             return prefixed_string
         return string
 
-    # Writes metadata information to log.
-    def writeInfo(self):
-        log.separator(
-            msg=(
-                'FINALIZED: ' + self.metadata.title +
-                ', ID: ' + self.metadata.id
-            ),
-            log_level="info"
-        )
-
-        # Log basic metadata
-        data_to_log = [
-            {'ASIN': self.metadata.id},
-            {'Album poster URL': self.thumb},
-            {'Album publisher': self.metadata.studio},
-            {'Album release date': str(self.metadata.originally_available_at)},
-            {'Album sort title': self.metadata.title_sort},
-            {'Album summary': self.metadata.summary},
-            {'Album title': self.metadata.title},
-        ]
-        log.metadata(data_to_log, log_level="info")
-
-        # Log basic metadata stored in arrays
-        multi_arr = [
-            {'Genres': self.metadata.genres},
-            {'Moods(Authors)': self.metadata.moods},
-            {'Styles(Narrators)': self.metadata.styles},
-        ]
-        log.metadata_arrs(multi_arr, log_level="info")
-
-        log.separator(log_level="info")
-
     # Remove extra description text from the title
     def simplify_title(self):
         # If the title ends with a series part, remove it
@@ -109,29 +251,23 @@ class AlbumUpdateTool:
         # If the title ends with "unabridged"/"abridged", with or without parenthesis
         # remove them; case insensitive
         album_title = re.sub(r" *\(?(un)?abridged\)?$", "",
-                                album_title, flags=re.IGNORECASE)
+                             album_title, flags=re.IGNORECASE)
         # Trim any leading/trailing spaces just in case
         album_title = album_title.strip()
 
         return album_title
 
 
-class ArtistUpdateTool:
-    UPDATE_URL = 'https://api.audnex.us/authors/'
-
-    def __init__(self, force, lang, media, metadata):
-        self.date = None
-        self.force = force
-        self.genres = None
-        self.lang = lang
-        self.media = media
-        self.metadata = metadata
-        self.thumb = ''
-
+class ArtistUpdateTool(UpdateTool):
     def parse_api_response(self, response):
         """
             Parses keys from API into helper variables if they exist.
         """
+        # Set empty variables
+        self.date = None
+        self.genres = None
+        self.thumb = ''
+
         if 'description' in response:
             self.description = response['description']
         if 'genres' in response:
@@ -140,34 +276,6 @@ class ArtistUpdateTool:
             self.name = response['name']
         if 'image' in response:
             self.thumb = response['image']
-
-    # Writes metadata information to log.
-    def writeInfo(self):
-        log.separator(
-            msg=(
-                'FINALIZED: ' + self.metadata.title +
-                ', ID: ' + self.metadata.id
-            ),
-            log_level="info"
-        )
-
-        # Log basic metadata
-        data_to_log = [
-            {'ASIN': self.metadata.id},
-            {'Author bio': self.metadata.summary},
-            {'Author name': self.metadata.title},
-            {'Author poster URL': self.thumb},
-            {'Author sort name': self.metadata.title_sort},
-        ]
-        log.metadata(data_to_log, log_level="info")
-
-        # Log basic metadata stored in arrays
-        multi_arr = [
-            {'Genres': self.metadata.genres},
-        ]
-        log.metadata_arrs(multi_arr, log_level="info")
-
-        log.separator(log_level="info")
 
 
 class TagTool:
@@ -201,7 +309,6 @@ class TagTool:
         """
         contributor_regex = '.+?(?= -)'
         if not self.helper.metadata.moods or self.helper.force:
-            self.helper.metadata.moods.clear()
             # Loop through authors to check if it has contributor wording
             for author in self.helper.author:
                 if not re.match(contributor_regex, author['name']):